# [Geneva.Logs] Support prefix-based table name mapping