--- a/src/OpenTelemetry.Exporter.Geneva/GenevaExporterOptions.cs
+++ b/src/OpenTelemetry.Exporter.Geneva/GenevaExporterOptions.cs
@@ -47,14 +47,32 @@ public class GenevaExporterOptions
 
             foreach (var entry in value)
             {
+                if (string.IsNullOrWhiteSpace(entry.Key))
+                {
+                    throw new ArgumentException("A table name mapping key was null, empty, or consisted only of white-space characters.", nameof(this.TableNameMappings));
+                }
+
                 if (string.IsNullOrWhiteSpace(entry.Value))
                 {
-                    throw new ArgumentException($"A string-typed value provided for {nameof(this.TableNameMappings)} must not be null, empty, or consist only of white-space characters.");
+                    throw new ArgumentException($"The table name mapping value provided for key '{entry.Key}' was null, empty, or consisted only of white-space characters.", nameof(this.TableNameMappings));
                 }
 
                 if (Encoding.UTF8.GetByteCount(entry.Value) != entry.Value.Length)
                 {
-                    throw new ArgumentException($"A string-typed value provided for {nameof(this.TableNameMappings)} must not contain non-ASCII characters.", entry.Value);
+                    throw new ArgumentException($"The table name mapping value '{entry.Value}' provided for key '{entry.Key}' contained non-ASCII characters.", nameof(this.TableNameMappings));
+                }
+
+                if (entry.Value != "*")
+                {
+                    if (!TableNameSerializer.IsValidTableName(entry.Value))
+                    {
+                        throw new ArgumentException($"The table name mapping value '{entry.Value}' provided for key '{entry.Key}' contained invalid characters or was too long.", nameof(this.TableNameMappings));
+                    }
+
+                    if (TableNameSerializer.IsReservedTableName(entry.Value))
+                    {
+                        throw new ArgumentException($"The table name mapping value '{entry.Value}' provided for key '{entry.Key}' is reserved and cannot be specified.", nameof(this.TableNameMappings));
+                    }
                 }
 
                 copy[entry.Key] = entry.Value;