--- a/src/OpenTelemetry.Exporter.Geneva/Internal/TableNameSerializer.cs
+++ b/src/OpenTelemetry.Exporter.Geneva/Internal/TableNameSerializer.cs
@@ -1,0 +1,313 @@
+// <copyright file="TableNameSerializer.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Text;
+
+namespace OpenTelemetry.Exporter.Geneva;
+
+internal sealed class TableNameSerializer
+{
+    public const int MaxSanitizedEventNameLength = 50;
+    public const int MaxCachedSanitizedTableNames = 1024;
+
+#pragma warning disable CA1825 // Avoid zero-length array allocations
+    /* Note: We don't use Array.Empty<byte> here because we need a different
+    instance to detect the pass-through case. */
+    private static readonly byte[] s_passthroughTableName = new byte[0];
+#pragma warning restore CA1825 // Avoid zero-length array allocations
+
+    private readonly byte[] m_defaultTableName;
+    private readonly Dictionary<string, byte[]> m_tableMappings;
+    private readonly bool m_shouldPassThruTableMappings;
+    private readonly object m_lockObject = new();
+    private TableNameCacheDictionary m_tableNameCache = new();
+
+    public ITableNameCacheDictionary TableNameCache => this.m_tableNameCache;
+
+    public TableNameSerializer(GenevaExporterOptions options, string defaultTableName)
+    {
+        Debug.Assert(options != null, "options were null");
+        Debug.Assert(!string.IsNullOrWhiteSpace(defaultTableName), "defaultEventName was null or whitespace");
+        Debug.Assert(IsValidTableName(defaultTableName), "defaultEventName was invalid");
+
+        this.m_defaultTableName = BuildStr8BufferForAsciiString(defaultTableName);
+
+        if (options.TableNameMappings != null)
+        {
+            var tempTableMappings = new Dictionary<string, byte[]>(options.TableNameMappings.Count, StringComparer.Ordinal);
+            foreach (var kv in options.TableNameMappings)
+            {
+                if (kv.Key == "*")
+                {
+                    if (kv.Value == "*")
+                    {
+                        this.m_shouldPassThruTableMappings = true;
+                    }
+                    else
+                    {
+                        this.m_defaultTableName = BuildStr8BufferForAsciiString(kv.Value);
+                    }
+                }
+                else if (kv.Value == "*")
+                {
+                    tempTableMappings[kv.Key] = s_passthroughTableName;
+                }
+                else
+                {
+                    tempTableMappings[kv.Key] = BuildStr8BufferForAsciiString(kv.Value);
+                }
+            }
+
+            this.m_tableMappings = tempTableMappings;
+        }
+    }
+
+    public static bool IsReservedTableName(string tableName)
+    {
+        Debug.Assert(!string.IsNullOrWhiteSpace(tableName), "tableName was null or whitespace");
+
+        // TODO: Implement this if needed.
+
+        return false;
+    }
+
+    public static bool IsValidTableName(string tableName)
+    {
+        Debug.Assert(!string.IsNullOrWhiteSpace(tableName), "tableName was null or whitespace");
+
+        var length = tableName.Length;
+        if (length > MaxSanitizedEventNameLength)
+        {
+            return false;
+        }
+
+        char firstChar = tableName[0];
+        if (firstChar < 'A' || firstChar > 'Z')
+        {
+            return false;
+        }
+
+        for (int i = 1; i < length; i++)
+        {
+            char cur = tableName[i];
+            if ((cur >= 'a' && cur <= 'z') || (cur >= 'A' && cur <= 'Z') || (cur >= '0' && cur <= '9'))
+            {
+                continue;
+            }
+
+            return false;
+        }
+
+        return true;
+    }
+
+    public int ResolveAndSerializeTableNameForCategoryName(byte[] destination, int offset, string categoryName, out ReadOnlySpan<byte> tableName)
+    {
+        byte[] mappedTableName = this.ResolveTableMappingForCategoryName(categoryName);
+
+        if (mappedTableName == s_passthroughTableName)
+        {
+            // Pass-through mode with a full cache.
+
+            int bytesWritten = SanitizedCategoryName(new Span<byte>(destination, offset, MaxSanitizedEventNameLength + 2), categoryName);
+
+            tableName = new ReadOnlySpan<byte>(destination, offset, bytesWritten);
+
+            return offset + bytesWritten;
+        }
+
+        tableName = mappedTableName;
+
+        return MessagePackSerializer.SerializeSpan(destination, offset, tableName);
+    }
+
+    private static byte[] BuildStr8BufferForAsciiString(string value)
+    {
+        var length = value.Length;
+
+        byte[] buffer = new byte[length + 2];
+
+        Encoding.ASCII.GetBytes(value, 0, length, buffer, 2);
+
+        MessagePackSerializer.WriteStr8Header(buffer, 0, length);
+
+        return buffer;
+    }
+
+    // This method would map the logger category to a table name which only contains alphanumeric values with the following additions:
+    // Any character that is not allowed will be removed.
+    // If the resulting string is longer than 50 characters, only the first 50 characters will be taken.
+    // If the first character in the resulting string is a lower-case alphabet, it will be converted to the corresponding upper-case.
+    // If the resulting string still does not comply with Rule, the category name will not be serialized.
+    private static int SanitizedCategoryName(Span<byte> buffer, string categoryName)
+    {
+        // Reserve 2 bytes for storing LIMIT_MAX_STR8_LENGTH_IN_BYTES and (byte)validNameLength -
+        // these 2 bytes will be back filled after iterating through categoryName.
+        int cursor = 2;
+        int validNameLength = 0;
+
+        // Special treatment for the first character.
+        var firstChar = categoryName[0];
+        if (firstChar >= 'A' && firstChar <= 'Z')
+        {
+            buffer[cursor++] = (byte)firstChar;
+            ++validNameLength;
+        }
+        else if (firstChar >= 'a' && firstChar <= 'z')
+        {
+            // If the first character in the resulting string is a lower-case alphabet,
+            // it will be converted to the corresponding upper-case.
+            buffer[cursor++] = (byte)(firstChar - 32);
+            ++validNameLength;
+        }
+        else
+        {
+            // Not a valid name.
+            return 0;
+        }
+
+        for (int i = 1; i < categoryName.Length; ++i)
+        {
+            if (validNameLength == MaxSanitizedEventNameLength)
+            {
+                break;
+            }
+
+            var cur = categoryName[i];
+            if ((cur >= 'a' && cur <= 'z') || (cur >= 'A' && cur <= 'Z') || (cur >= '0' && cur <= '9'))
+            {
+                buffer[cursor++] = (byte)cur;
+                ++validNameLength;
+            }
+        }
+
+        // Backfilling MessagePack serialization protocol and valid category length to the startIdx of the categoryName byte array.
+        MessagePackSerializer.WriteStr8Header(buffer, 0, validNameLength);
+
+        return cursor;
+    }
+
+    private byte[] ResolveTableMappingForCategoryName(string categoryName)
+    {
+        var tableNameCache = this.m_tableNameCache;
+
+        if (tableNameCache.TryGetValue(categoryName, out byte[] tableName))
+        {
+            return tableName;
+        }
+
+        return this.ResolveTableMappingForCategoryNameRare(categoryName);
+    }
+
+    private byte[] ResolveTableMappingForCategoryNameRare(string categoryName)
+    {
+        lock (this.m_lockObject)
+        {
+            var tableNameCache = this.m_tableNameCache;
+
+            // Check if another thread added the mapping while we waited on the
+            // lock.
+            if (tableNameCache.TryGetValue(categoryName, out byte[] tableName))
+            {
+                return tableName;
+            }
+
+            byte[] mappedTableName = null;
+
+            // If user configured table name mappings run resolution logic.
+            if (this.m_tableMappings != null
+                && !this.m_tableMappings.TryGetValue(categoryName, out mappedTableName))
+            {
+                // Find best match if an exact match was not found.
+
+                string currentKey = null;
+
+                foreach (var mapping in this.m_tableMappings)
+                {
+                    if (!categoryName.StartsWith(mapping.Key, StringComparison.OrdinalIgnoreCase))
+                    {
+                        continue;
+                    }
+
+                    if (currentKey == null || mapping.Key.Length >= currentKey.Length)
+                    {
+                        currentKey = mapping.Key;
+                        mappedTableName = mapping.Value;
+                    }
+                }
+            }
+
+            mappedTableName ??= !this.m_shouldPassThruTableMappings
+                ? this.m_defaultTableName
+                : s_passthroughTableName;
+
+            if (mappedTableName.Length <= 0 && tableNameCache.CachedSanitizedTableNameCount < MaxCachedSanitizedTableNames)
+            {
+                // We resolved to a wildcard which is pass-through mode.
+
+                Span<byte> stackBuffer = stackalloc byte[64];
+
+                int bytesWritten = SanitizedCategoryName(stackBuffer, categoryName);
+                if (bytesWritten > 0)
+                {
+                    mappedTableName = stackBuffer.Slice(0, bytesWritten).ToArray();
+
+                    tableNameCache.CachedSanitizedTableNameCount++;
+                }
+                else
+                {
+                    // Note: When the table name could not be sanitized we cache
+                    // the empty array NOT s_passthroughTableName.
+                    mappedTableName = Array.Empty<byte>();
+                }
+            }
+
+            // Note: This is using copy-on-write pattern to keep the happy
+            // path lockless once everything has spun up.
+            this.m_tableNameCache = new(tableNameCache)
+            {
+                [categoryName] = mappedTableName,
+            };
+
+            return mappedTableName;
+        }
+    }
+
+    // Note: This is used for tests.
+    public interface ITableNameCacheDictionary : IReadOnlyDictionary<string, byte[]>
+    {
+        int CachedSanitizedTableNameCount { get; }
+    }
+
+    private sealed class TableNameCacheDictionary : Dictionary<string, byte[]>, ITableNameCacheDictionary
+    {
+        public TableNameCacheDictionary()
+            : base(0, StringComparer.OrdinalIgnoreCase)
+        {
+        }
+
+        public TableNameCacheDictionary(TableNameCacheDictionary sourceCache)
+            : base(sourceCache, StringComparer.OrdinalIgnoreCase)
+        {
+            this.CachedSanitizedTableNameCount = sourceCache.CachedSanitizedTableNameCount;
+        }
+
+        public int CachedSanitizedTableNameCount { get; set; }
+    }
+}