## Changes

This PR adds support for prefix-based resolution (similar to [logging configuration](https://learn.microsoft.com/aspnet/core/fundamentals/logging/#configure-logging)) to the table mapping configuration options.

## Usage

Given this config...

```csharp
        var options = new GenevaExporterOptions
        {
            TableNameMappings = new Dictionary<string, string>()
            {
                ["*"] = "DefaultLogs",
                ["MyCompany"] = "InternalLogs",
                ["MyCompany.Product1"] = "InternalProduct1Logs",
                ["MyCompany.Product2"] = "InternalProduct2Logs",
                ["MyCompany.Product2.Security"] = "InternalSecurityLogs",
                ["MyPartner"] = "*",
            },
        };
```

* `ILogger<ThirdParty.Thing>`: This would go to "DefaultLogs"
* `ILogger<MyCompany.ProductX.Thing>`: This would go to "InternalLogs"
* `ILogger<MyCompany.Product1.Thing>`: This would go to "InternalProduct1Logs"
* `ILogger<MyCompany.Product2.Thing>`: This would go to "InternalProduct2Logs"
* `ILogger<MyCompany.Product2.Security.Alert>`: This would go to "InternalSecurityLogs"
* `ILogger<MyPartner.Product.Thing>`: This is marked as pass-through ("*") so it will be sanitized as "MyPartnerProductThing" table name

## Implementation details

The implementation is close to what ILogger does for its filter configuration. The idea is you don't want to have to apply the rules each time you process a log message because the perf would be terrible. You want to have the [result of rule selection](https://github.com/dotnet/runtime/blob/97f51dcca8e3e7ce85cbe52182db922f5afb90b0/src/libraries/Microsoft.Extensions.Logging/src/LoggerRuleSelector.cs#L10) computed once and available each time a log message is written. What the logging framework does is store the result on the `ILogger` instance created/injected from the factory. We don't have that luxury here. What I did was introduce a [helper class which maintains a cache](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/796/files#diff-0f16ceb5356860833a44dc4c125381d9560652433f15b48a504d6ad90f217209R39) of the final category -> table name. It is backed by a dictionary so you pay for a hash on the lookup which is what we were paying before (if mappings were used at all).

Some details about the cache...

* Lookups are lock-free. If a new mapping is needed we do a copy-on-write. So warmup is expensive but once you are warm it is cheap.

* There is no constraint on the number of keys (category names) which will be cached. ILogger itself maintains an unconstrained cache of loggers by category name so it seemed like that would OOM well before this.

* If the mapping is pass-through we only cache up to 1024 of the sanitized table names. Once the max is reached we just write them out to the buffer as they are logged (this was the existing behavior always).

## Benchmarks

Implementation + optimizations actually speed things up for everyone.

Before:
|                    Method | IncludeFormattedMessage |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|-------------------------- |------------------------ |---------:|---------:|---------:|-------:|----------:|
| LoggerWithMessageTemplate |                   False | 546.6 ns | 10.43 ns |  9.76 ns | 0.0162 |     208 B |
| LoggerWithDirectLoggerAPI |                   False | 509.0 ns |  9.98 ns | 11.09 ns | 0.0267 |     344 B |
| LoggerWithSourceGenerator |                   False | 529.5 ns |  5.30 ns |  4.43 ns | 0.0134 |     168 B |
|        SerializeLogRecord |                   False | 359.4 ns |  5.70 ns |  5.34 ns | 0.0081 |     104 B |
|                    Export |                   False | 390.7 ns |  5.24 ns |  4.90 ns | 0.0081 |     104 B |
| LoggerWithMessageTemplate |                    True | 588.1 ns | 10.01 ns |  9.36 ns | 0.0162 |     208 B |
| LoggerWithDirectLoggerAPI |                    True | 512.0 ns |  7.44 ns |  6.96 ns | 0.0267 |     344 B |
| LoggerWithSourceGenerator |                    True | 537.0 ns |  8.15 ns |  6.80 ns | 0.0134 |     168 B |
|        SerializeLogRecord |                    True | 361.9 ns |  5.39 ns |  5.05 ns | 0.0081 |     104 B |
|                    Export |                    True | 391.7 ns |  4.14 ns |  3.87 ns | 0.0081 |     104 B |

After:
|                    Method | IncludeFormattedMessage |     Mean |    Error |  StdDev |   Gen0 | Allocated |
|-------------------------- |------------------------ |---------:|---------:|--------:|-------:|----------:|
| LoggerWithMessageTemplate |                   False | 530.6 ns |  3.30 ns | 2.92 ns | 0.0162 |     208 B |
| LoggerWithDirectLoggerAPI |                   False | 496.9 ns |  8.87 ns | 8.30 ns | 0.0267 |     344 B |
| LoggerWithSourceGenerator |                   False | 516.1 ns |  8.12 ns | 7.60 ns | 0.0134 |     168 B |
|        SerializeLogRecord |                   False | 343.4 ns |  6.64 ns | 6.52 ns | 0.0081 |     104 B |
|                    Export |                   False | 383.6 ns |  5.77 ns | 5.39 ns | 0.0081 |     104 B |
| LoggerWithMessageTemplate |                    True | 566.3 ns | 10.16 ns | 9.50 ns | 0.0162 |     208 B |
| LoggerWithDirectLoggerAPI |                    True | 504.9 ns |  9.36 ns | 8.29 ns | 0.0267 |     344 B |
| LoggerWithSourceGenerator |                    True | 524.2 ns |  5.07 ns | 4.75 ns | 0.0134 |     168 B |
|        SerializeLogRecord |                    True | 349.7 ns |  4.56 ns | 4.26 ns | 0.0081 |     104 B |
|                    Export |                    True | 380.9 ns |  3.93 ns | 3.67 ns | 0.0081 |     104 B |

## TODOs

* [X] Appropriate `CHANGELOG.md` updated for non-trivial changes
* [X] Unit tests
* [X] README updates